Struct liquery_file::QueryFile

pub struct QueryFile {
    // some fields omitted
}

A Queryable instance of a file

Methods

impl QueryFile

Query the file for some information

Queryable fields

size - returns size of file in bytes

filetype - returns file if path points to file, directory if path points to directory and symlink if path points to a symlink.

extension - returns file extension, see Path::extension() for more details.

filename - returns the file name

mimetype - returns a guess of the mimetype based off the file name

fn new<P: AsRef<Path>>(path: P) -> Result<Self>

Get a new instance of a QueryFile

Example

use liquery_file::QueryFile;

let queryable = QueryFile::new("file.txt");